Buat Models dan file todo_model
