+++ lib/controllers/todo_controller.dart
@@ -1,18 +1,5 @@
 import 'package:get/get.dart';
-
-class Todo {
-  String title;
-  String description;
-  String category;
-  bool isDone;
-
-  Todo({
-    required this.title,
-    required this.description,
-    required this.category,
-    this.isDone = false,
-  });
-}
+import 'package:todolistapp_27_26/models/todo_model.dart';
 
 class TodoController extends GetxController {
   var todos = <Todo>[].obs;
