+++ lib/models/todo_model.dart
@@ -1,0 +1,13 @@
+class Todo {
+  String title;
+  String description;
+  String category;
+  bool isDone;
+
+  Todo({
+    required this.title,
+    required this.description,
+    required this.category,
+    this.isDone = false,
+  });
+}
