+++ lib/pages/dashboard_page.dart
@@ -1,32 +1,36 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:todolistapp_27_26/controllers/dashboard_controller.dart';
 import 'package:todolistapp_27_26/controllers/home_controller.dart';
 import 'home_page.dart';
 import 'history_page.dart';
 import 'profile_page.dart';
 
 class DashboardPage extends StatelessWidget {
-  final HomeController homeController = Get.find<HomeController>();
-
-  final List<Widget> pages = [
-    HomePage(),
-    HistoryPage(),
-    ProfilePage(),
-  ];
+  final dashboardController = Get.put(DashboardController());
 
   @override
   Widget build(BuildContext context) {
-    return Obx(() => Scaffold(
-          body: pages[homeController.currentIndex.value],
-          bottomNavigationBar: BottomNavigationBar(
-            currentIndex: homeController.currentIndex.value,
-            onTap: (index) => homeController.changePage(index),
-            items: const [
-              BottomNavigationBarItem(icon: Icon(Icons.home), label: "Home"),
-              BottomNavigationBarItem(icon: Icon(Icons.history), label: "History"),
-              BottomNavigationBarItem(icon: Icon(Icons.person), label: "Profile"),
-            ],
-          ),
-        ));
+    return Obx(
+      () => Scaffold(
+        body: dashboardController.pages[dashboardController.currentIndex.value],
+        bottomNavigationBar: BottomNavigationBar(
+          currentIndex: dashboardController.currentIndex.value,
+          onTap: (index) => dashboardController.changePage(index),
+          items: const [
+            BottomNavigationBarItem(
+             icon: Icon(Icons.home),
+             label: "Home"),
+            BottomNavigationBarItem(
+              icon: Icon(Icons.history),
+              label: "History",
+            ),
+            BottomNavigationBarItem(
+             icon: Icon(Icons.person),
+             label: "Profile"),
+          ],
+        ),
+      ),
+    );
   }
 }
