+++ lib/widgets/todo_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:todolistapp_27_26/controllers/todo_controller.dart';
-
+import 'package:todolistapp_27_26/models/todo_model.dart';
 class TodoCard extends StatelessWidget {
   final Todo todo;
   final VoidCallback? onDone; // optional (nullable)
